OSDC::Israel 2007 has been cancelled, or at the very least postponed until much later in the year.  Therefore I have removed it from the overall Workshop list, but I have noted this status change on the front page so as to not confuse people ('Hey, where did that go?').
This is the mail from Gabor Szabo about it, sent on 3 December:

I am sorry to announce that the Open Source Developers' Conference
planned to be in February 2007 will not take place.

Even though we started the process early this year, due to lack of
time I was not able to
finish the management software on time and to do all the organizational
things.

Therefore I have no cancel or at least postpone the event.

Later on I'll look at my schedule to see if it is possible to do it in
June-July 2007
or if we can only do it again in February 2008.
Your opinion is welcome here.


Next to the conference we planned to do a Pugs Hackathon. As it
requires much less effort to organize that I am quite sure we will
have a 1 week long Pugs Hackathon sometime
around end of February, beginning of March with Audrey Tang.

If you are interested to participate (even for 2-3 days) please let me know.


git-svn-id: svn://svn.mongueurs.net/yef-www/trunk@92 c065baed-db06-db11-a765-00306e02d86a