I have started converting workshops to a data structure similar to what we are using for news and conferences.  As with conferences, this will allow us to post some information on the main page (I was thinking up to 3 conferences which haven't yet taken place) without duplicating data.
Note that I only put in 3 conferences from last year (adding the London
Workshop, which wasn't on the original list) and none from the previous
years yet.  I will get around to this at some point, or someone else
can do it.  It's just a matter of restructuring the information on the
present Workshops page in to the data structure found in LIB/workshop_items,
and looking up the city and the language from the workshop's website.


git-svn-id: svn://svn.mongueurs.net/yef-www/branches/restructure-2006-branch@167 c065baed-db06-db11-a765-00306e02d86a