Workshops from 2005 have been added.
Some URLs overall were modified, to point to older conferences rather than
the most recent conference in the region.

I also added all the Hungarian Perl workshops, including the ones which
weren't listed on our page originally, because there was no reason not to
include them (I had the page to the workshop open and found the additional
information there).


git-svn-id: svn://svn.mongueurs.net/yef-www/branches/restructure-2006-branch@191 c065baed-db06-db11-a765-00306e02d86a