Replaced 'finance' with 'sitemap' in the navigation.
I think finance is very important, but it is a part of the foundation (so
is 'donating', but we leave it in the menu to make it very easy for people
to find).  I removed it because I felt it was subordinate and the navigation
was getting too full.

I added the sitemap because I think the site is now getting rather 
complicated, with many pages not accessible from the navigation but only
by following a link from a page, and I want to make it possible for people
to jump to different areas of the site without knowing where the links are.



git-svn-id: svn://svn.mongueurs.net/yef-www/branches/lisbon2009@716 c065baed-db06-db11-a765-00306e02d86a